Refactor into packages
There are several classes which are not organised into packages.

Refactoring these classes into different packages improves code structure.

As a step towards such unification, we can organise common classes into packages. Doing so will make it easier for readers to understand the groups of classes within the code.
