+++ src/main/java/duke/command/FindCommand.java
@@ -1,8 +1,8 @@
 package duke.command;
 
-import duke.CommandResult;
-import duke.DukeException;
-import duke.TaskList;
+import duke.commandresult.CommandResult;
+import duke.exception.DukeException;
+import duke.tasklist.TaskList;
 
 public class FindCommand extends Command {
