+++ src/main/java/duke/command/ListNumberPrintable.java
@@ -1,6 +1,6 @@
 package duke.command;
 
-import duke.TaskList;
+import duke.tasklist.TaskList;
 
 /**
  * This interface allows a contract to be created to return the string of a given TaskList.
