+++ src/main/java/duke/command/TaskListAddable.java
@@ -1,7 +1,7 @@
 package duke.command;
 
 import duke.task.Task;
-import duke.TaskList;
+import duke.tasklist.TaskList;
 
 /**
  * This interface allows a contract to be created to return the result of adding a task to a TaskList.
