+++ src/main/java/duke/commandresult/CommandResult.java
@@ -1,4 +1,4 @@
-package duke;
+package duke.commandresult;
 
 public class CommandResult {
     private final String feedbackResult;
