+++ src/main/java/duke/storage/Storage.java
@@ -1,4 +1,8 @@
-package duke;
+package duke.storage;
+
+import duke.exception.DukeException;
+import duke.formatter.Formatter;
+import duke.tasklist.TaskList;
 
 import java.io.BufferedWriter;
 import java.io.BufferedReader;
