+++ src/main/java/duke/task/Deadline.java
@@ -1,6 +1,6 @@
 package duke.task;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 
 /**
  * The Deadline type task.
