+++ src/main/java/duke/task/Event.java
@@ -1,6 +1,6 @@
 package duke.task;
 
-import duke.DukeException;
+import duke.exception.DukeException;
 
 /**
  * The Event type task.
