+++ src/main/java/duke/task/TimedTask.java
@@ -1,7 +1,7 @@
 package duke.task;
 
-import duke.DateTimeFormatCreator;
-import duke.DukeException;
+import duke.formatter.DateTimeFormatCreator;
+import duke.exception.DukeException;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
